Fixed hardpoints for moors
I moved the 4 moor HpDockMounts away 1 meter and then lined all of the HpDockPoints so that docking and launching are perfectly straight. All HpDockMounts and HpDockPoints are the same distance from their origin points.

I changed the angles for Dock H and I to 70 degrees instead of 65. Much more accurate now.