Fixing empathy errors in Kusari factions
GMG in the status quo is openly in conflict for territory with the Corsairs, after Corsairs and Hogosha have betrayed the guild in the Artifact deal. They offer missions against corsairs on all 3 of their bases.

Samura is the literal patron of the two criminal factions that are openly hostile to foreign corporations and "accidentally" raid Kishiro cargo as well. Currently, the Hogosha are assigned backwards and the FA reacts positively to eliminating Samura ships.