Remove Cloak Activator fire-cancelling
The cloak activator is being used on-fire for triggering energy drop.
If the ship cloaks in, the weapons were all cancelled on server-side due this.
It would be just correct to let the server draw even the activator's energy now.
A downside is potential zero-second existing ammo of the activator.