Increase spacing between missile alert and NN icon window in MP
Previously the missile alert and NN icon windows would overlap if you played on Multiplayer. This happened because the icon window has two additional icons in MP, making it wider. Thus the right-side alerts have been shifted more towards the center of the screen. The same has been done for the left-side alerts to ensure symmetry.