New SUR for the updated Rhino Li_freighter
New collision surface for the Rhino with updated HP boxes. Built to Maxx 0.99. Utilizes the shield bubble from Discos version just cuz. Weapon Points are now normal size as the SUR is form fitting and they dont have to poke out of the amorphous  blob of the vanilla sur to be actually useful.